// DotsPasswordManager-client/src/app/core/main-api/models/user-saved-password-dt-os-saved-password-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserSavedPasswordDtOsSavedPasswordDto {
  CreatedAt?: string;
  IsFavourite?: boolean;
  Login: string;
  Name: string;
  Notes?: string;
  Password: string;
  PasswordId?: string;
  SecondLogin?: string | null;
  Tags?: Array<string>;
  UpdatedAt?: string;
  Url?: string;
}